Allow to specify URL instead of HTML

+++ index.js
@@ -9,14 +9,19 @@ var quote   = require('shell-quote').quote;
 app.use(express.json());
 
 app.post('/', function(req, res) {
-  var options = _.map(_.omit(req.body, 'html'), function(value, name) {
+  var options = _.map(_.omit(req.body, 'html', 'url'), function(value, name) {
     return '--' + name.replace(/_/g, '-') + ' ' + quote([_.toString(value)]);
   }).join(' ');
 
-  var path = temp.sync(_.toString(req.body.html), 'index.html');
-  setTimeout(function() { fs.unlink(path); }, 1000 * 60 * 60); // One hour.
+  if (!_.isEmpty(req.body.url)) {
+    var url = req.body.url;
+  } else {
+    var path = temp.sync(_.toString(req.body.html), 'index.html');
+    var url  = 'file://' + path;
+    setTimeout(function() { fs.unlink(path); }, 1000 * 60 * 60); // One hour.
+  }
 
-  var command = 'wkhtmltopdf ' + options + ' file://' + path + ' - | cat';
+  var command = 'wkhtmltopdf ' + options + ' ' + url + ' - | cat';
 
   console.log('Executing ' + command);
